Extract PackageSourceTests from GenericSourceTests
