+++ platform-tests/src/test/java/org/junit/platform/engine/support/descriptor/GenericSourceTests.java
@@ -11,7 +11,6 @@
 package org.junit.platform.engine.support.descriptor;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import java.io.Serializable;
 import java.lang.reflect.Method;
@@ -19,10 +18,9 @@ import java.util.stream.Stream;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInfo;
-import org.junit.platform.commons.util.PreconditionViolationException;
 
 /**
- * Unit tests for {@link PackageSource} and {@link MethodSource}.
+ * Unit tests for {@link MethodSource}.
  *
  * @since 1.0
  */
@@ -31,41 +29,9 @@ class GenericSourceTests extends AbstractTestSourceTests {
 	@Override
 	Stream<Serializable> createSerializableInstances() throws Exception {
 		return Stream.of( //
-			PackageSource.from("package.source"), //
 			MethodSource.from(getMethod("method1")), //
 			MethodSource.from(getMethod("method2")) //
 		);
-	}
-
-	@Test
-	void packageSourceFromNullPackageName() {
-		assertThrows(PreconditionViolationException.class, () -> PackageSource.from((String) null));
-	}
-
-	@Test
-	void packageSourceFromEmptyPackageName() {
-		assertThrows(PreconditionViolationException.class, () -> PackageSource.from("  "));
-	}
-
-	@Test
-	void packageSourceFromPackageName() {
-		String testPackage = getClass().getPackage().getName();
-		PackageSource source = PackageSource.from(testPackage);
-
-		assertThat(source.getPackageName()).isEqualTo(testPackage);
-	}
-
-	@Test
-	void packageSourceFromNullPackageReference() {
-		assertThrows(PreconditionViolationException.class, () -> PackageSource.from((Package) null));
-	}
-
-	@Test
-	void packageSourceFromPackageReference() {
-		Package testPackage = getClass().getPackage();
-		PackageSource source = PackageSource.from(testPackage);
-
-		assertThat(source.getPackageName()).isEqualTo(testPackage.getName());
 	}
 
 	@Test
@@ -76,13 +42,6 @@ class GenericSourceTests extends AbstractTestSourceTests {
 		assertThat(source.getClassName()).isEqualTo(getClass().getName());
 		assertThat(source.getMethodName()).isEqualTo(testMethod.getName());
 		assertThat(source.getMethodParameterTypes()).isEqualTo(String.class.getName());
-	}
-
-	@Test
-	void equalsAndHashCodeForPackageSource() {
-		Package pkg1 = getClass().getPackage();
-		Package pkg2 = String.class.getPackage();
-		assertEqualsAndHashCode(PackageSource.from(pkg1), PackageSource.from(pkg1), PackageSource.from(pkg2));
 	}
 
 	@Test
